+++ l10_mapper_annotation/lib/l10_mapper_annotation.dart
@@ -1,3 +1,0 @@
-library l10_mapper_annotation;
-
-export 'src/l10_mapper_annotation.dart';