+++ l10_mapper_annotation/lib/l10n_mapper_annotation.dart
@@ -1,0 +1,3 @@
+library l10n_mapper_annotation;
+
+export 'src/l10n_mapper_annotation.dart';